Remove FragmentClassSet from BuildConfigurationKey
Ultimately, this feature was not being used and thus contributing to technical debt. In particular, note how the background on constructing ConfiguredTargetKey can now be simplified as no longer need to consider if a given BuildConfigurationKey is 'canonical'.

Future use sites that want to trim based on fragments and their associated option requirements can just re-implement the trimming logic. This could possibly be escalated to a new, separate SkyFunction; potentially also incorporating running (and thus more centrally caching) transition application.

PiperOrigin-RevId: 410362732