Enforce that JavaCompileAction.outputDepsProto is non-null if classPathMode is BAZEL: we unconditionally read the proto output in such cases.
Also skip a useless read of the jdeps file when doing full classpath compilation.

This is in preparation for threading "mandatory outputs" through to remote execution so that it can eagerly fail remote actions that don't produce their required outputs. outputDepsProto is the only required output on the initial "reduced" attempt, but it would be nice to know it's not null.

PiperOrigin-RevId: 414850626