Allow errors to be printed in a no-keep-going build after an action is interrupted: if the action was interrupted by a file-level error during execution, suppressing further errors means that the user never gets to see the errors constructed during error bubbling.
This does leave the door open to a user manually interrupting the build and then seeing an error from some corner of Blaze (or a local action?) that doesn't react well to being interrupted. We should just fix those if they are found.

This bug was introduced in https://github.com/bazelbuild/bazel/commit/fd2a779687616bbfb18b5110cbf708a24a681f7a, which put ActionExecutionFunction in charge of deciding when to stop printing errors.

PiperOrigin-RevId: 416593263