Add option to start locally scheduling actions above a certain age in dynamic scheduling. This allows handling when remote systems get stuck for whatever reason.
Currently, local actions are scheduled LIFO, so we have the best chance of actually doing something faster locally, rather than getting cancelled when the remote branch finishes shortly after. With this change, we switch to FIFO for actions that are above a certain age, where we guess the remote is not going to answer in a reasonable time.

We can't tell what timeouts to expect, since they could be caused by many different parts of the remote system. Instead, this allows adjusting based on experience. If only very few remote calls end up timing out, setting this to ~half the seen timeout should be fine. The more often remote calls time out, the lower you want to set this.

Setting this flag can hide problems in your remote system. Make sure to have proper monitoring and SLOs in place for your remote system before papering it over with this flag.

PiperOrigin-RevId: 422434741