Remove one extra Skyframe evaluation to get the Targets for testing.
Getting the Target objects from the package manager costs 1 Skyframe evaluation
per target. We already have this ready in an earlier stage of the build, so
there's no need to do it again when collecting the tests.

Since the values are likely already cached in Skyframe by the time we call this, no performance gain is expected.

PiperOrigin-RevId: 424070958