Pass rule attributes to compile_action via parameters.
The change strives to make parts of Java rules more composable, that is compile_action becomes reusable from other rules. (Along the way potentially improving also documentation and testability).

Computing parameter values based on ctx had to be moved into java_common call. Also computing runfiles was moved to java_common (because runtime_deps and exports needed value of DefaultInfo provider and not just JavaInfo).

The compile_action function becomes rather small and most of the logic is implemented in java_common.compile. I don't have a particular worries about that as with further Starlarkfication the function will grow.

There are test added/changed, because the cl is a functionally no-op.

output_prefix was used for implicit outputs workarounds and was not used.

PiperOrigin-RevId: 424345208