Use a class hierarchy for `PathFragment` to encode the information formerly encoded in `PathFragment#driveStrLength`.
By doing this, we reduce the shallow heap of a `PathFragment` instance (actually, an instance of one of the three new subclasses) from 24 bytes to 16 bytes. Since Bazel typical has many instances retained in memory (think: "hundreds of thousands" or even "millions"), this saves a significant amount of memory.

Brief history of the codebase:
* Before unknown commit, the `PathFragment` class was unoptimized in this way.
* unknown commit introduced a similar optimization.
* unknown commit redesigned and reimplemented `PathFragment` with the goal of storing flat `String`s for the entire logical path string. We lost the memory optimization from unknown commit there.

PiperOrigin-RevId: 427354168