Open source `LocalDiffAwarenessTest.java`.
This doesn't work at all Mac machines, probably due to the same underlying issue as in https://github.com/bazelbuild/bazel/issues/10776. I added a comment in `LocalDiffAwarenessTest.java` documenting this.

This almost works on Windows machines, but there seems to be some non-determinism with the implementation of the feature on Windows. I tagged the test with `no_windows` and wrote a comment mentioning the issues I ran into.

PiperOrigin-RevId: 427445489