Don't intern any `FileStateValue.Key` instances.
The practical uses of `FileStateValue#key` mean that interning is unproductive in almost all cases, meaning we're spending memory on the interner's data structure without benefiting from it. Instead, interning _nothing_ is a fairly large net win (saves ~0.15% analysis-phase heap in my benchmarks). The reasoning for this is given in a large code comment.

Also, since we're not interning `FileStateValue.Key` it's now inefficiency wrt shallow heap for `FileStateValue.Key` to extend from `AbstractSkyKey` since we pay for the `AbstractSkyKey#hashCode` field but, per the above paragraph and also the lack of interning, the motivation in the comment for `AbstractSkyKey#hashCode` doesn't apply. So this CL additionally reduces the shallow heap of `FileStateValue.Key` from 24 bytes to 16 bytes.

PiperOrigin-RevId: 427593902