Stop passing dependencies' descriptor sets when generating descriptor sets
When we generate descriptor sets, we directly parse all .proto files in the
transitive dependency tree, so as far as I can tell there is no need to make
the dependency descriptor sets available. This change should speed up builds by
allowing much more parallelism, since it will be possible to generate the
descriptor set for a given proto file without having to wait for that step to
complete in all dependencies first.

RELNOTES: None.
PiperOrigin-RevId: 429446795