Abort query evaluation early in no-keep-going mode if preloading the transitive closure of a set of target fails *and* the transitive closure needed was unbounded.
This prevents pathological behavior of a no-keep-going query encountering an error,  printing it, aborting Skyframe preloading evaluation, and then sequentially loading packages in LabelVisitor because they weren't loaded in the preloading phase.

That pathological behavior is still possible with a depth bound, but this gets rid of a class of such issues.

To facilitate this, moved QueryTransitivePackagePreloader into the query2 tree, and made the maxDepth an OptionalInt. This might also theoretically allow us to process graphs of depth > Integer.MAX_VALUE although I'm pretty sure we'd blow up for a dozen other reasons first.

PiperOrigin-RevId: 429454209