Have README.md in java_tools release reference the right target
I was trying to figure out what produced the files included in https://github.com/bazelbuild/java_tools/releases/download/java_v11.6/java_tools-v11.6.zip that happens to have some instructions on how to re-produce the contents of the file.

However, it turns out that the instructions in README.md references the wrong build target, which made it a bit confusing. This change updates the target to one that produces a zip that seem to contain the right files.

Closes #15093.

PiperOrigin-RevId: 440331100