## Add support for Windows `linkstatic` computed default
This uses the same technique as `cc_binary` with/without aspects.
Essentially we are declaring two rules named cc_test with different private attribute defaults. This is primarily done this way to hide the implementation details of computed defaults until something like
[bazel/issues/14434](https://github.com/bazelbuild/bazel/issues/14434) is implemented.

NOTE: When `cc_test` written in Starlark is made the default, this _will_ introduce a behavior change -- notably that we will default linkstatic based on whether we are _targeting_ Windows, not whether the host Platform is Windows (as is done today).
The latter heuristic may have been the best available option before Platforms, but it does not match the actual intent.
PiperOrigin-RevId: 441472088