Use an event-based system to collect analyzed/built targets/tests/aspects.
This change puts in place the necessary classes for such a system, and apply said system to Skymeld (also experimental). A follow-up change will introduce this to the regular code path, guarded behind a flag.

There are basically 5 events:
- target analyzed
- test analyzed
- aspect analyzed
- target built
- aspect built

Each event is posted at most once. For Skymeld, the posting takes place from within a SkyFunction, that means we need to prevent posting multiple times when 1) the function restarts (with the use of State) and 2) error bubbling (by checking with the environment if we're in error bubbling).

Also changed the various calls to eventBus.post(...) in SkyFunctions to instead use env.getListener().post(...), which is a more canonical way to post events from Skyframe.

PiperOrigin-RevId: 444256098