Check status of write operations for `pid_file` in `daemonize` wrapper.
The `daemonize` tool writes the pid of spawned process into `pid_file` without checking the status. This, mixed with unchecked read in Bazel, results with a crash when we start from a full disk. Add checks around the write in `daemonize` and read in Bazel to make sure we get the pid successfully.

PiperOrigin-RevId: 444592395