Fix a narrow non-determinism bug in package loading.
The symptom of the bug is that the location assigned to input file targets Blaze implicitly creates is non-deterministic among the rule instance targets that specify the input file. The correct behavior is to choose the location of the first such rule instance target (wrt the order in which they were instantiated).

This non-determinism was introduced in https://github.com/bazelbuild/bazel/commit/736707833d94613e5da8e4adc32e8146bc4bc307. This CL here reverts part of that CL.

In order to fix this cleanly I also had to fix an existing bug with the hacky `#replaceTarget` (used only by `WorkspaceFileFunction`, which implements Bazel's `WORKSPACE` feature) that didn't matter in practice because the old `#getRules` implementation dodged it. The issue here was caught by existing tests, so that's nice at least.

PiperOrigin-RevId: 445489517