Don't skip filtering when manifest of source-file paths is empty
When a coverage entry doesn't correspond to a file listed in execPathsOfUninstrumentedFiles, it's discarded. However, since this class has a constructor which takes no arguments and sets execPathsOfUninstrumentedFiles to an empty-set. But this leads to those same no-filtering behavior when the constructor is passed an empty list of valid source files.

Instead, the "leave as is" behavior should be conditioned on `execPathsOfUninstrumentedFiles == null`, and an empty set should be treated consistently with other sets with no matching items.

RELNOTES: None.
PiperOrigin-RevId: 445497757