Point symlinks to headers with stripped prefix to exec root instead of source location.
Use the exec root location as the target for symlinks created for headers with stripped prefix to make sure the paths stay in sync with exec root, which could change across builds. If the header itself stays the same, then the symlink will not be updated, leading to different resolved path for the 2 paths.

PiperOrigin-RevId: 446709489