Remove dead code.
SkyframeAnalysisAndExecutionResult doesn't have to override #withAdditionalErroredTargets: it's only used to filter out targets that are incompatible with the platform/environment.

For Skymeld, such a case is handled "natively" within Skyframe: if a target is incompatible and warrants an error, the TargetCompatibilityCheckException would be thrown and handled like a regular exception, and the incompatible target would not be executed.

PiperOrigin-RevId: 446934424