Fix some bugs in SkyframeErrorProcessor.
1. Only send AnalysisFailureEvent in SkyframeErrorProcessor if the error is
analysis-related.
AnalysisFailureEvent tells us that the analysis of a top level target failed
and it wasn't executed at all. It's wrong to also send this if the error is
execution-related.

2. Handle an execution error originating from an Aspect.
This was an oversight when we introduced this combined mode of error handling for Skyframe. Added a test for this scenario.

PiperOrigin-RevId: 446950178