Remote: Prefetch input files into a temporary path first.
When building with build without bytes and dynamic execution, we need prefetch input files for local actions. Sometimes, multiple local actions could share the same input files, so there could be a case where multiple call sites share the same download instance. If the local action is cancelled (due to remote branch wins), the download it requested should also be cancelled only if that download is not shared with other local action (or all the releated local actions are cancelled).

Before this change, the inputs are written to their final destination directly. This is fine if we can make sure no race or bug in the prefetcher. However, this is not true: https://github.com/bazelbuild/bazel/issues/15010.

The root cause is, when cancelling the downloads, sometimes, the partially downloaded files on the disk are not deleted.

By making the prefetcher download input to a temporary path first, we can:
  1. Mitigate the race: only the final move step will potentially cause the race condition.
  2. Provide a way to observe the race: if these is no race, all temporary files should be either moved or deleted. But when running with this change, many temporary files exist.

Working towards https://github.com/bazelbuild/bazel/issues/12454.

PiperOrigin-RevId: 447473693