Add support for param files for Android NDK
Fix #7982

Currently, on Windows, `ar` command can fail when using Android NDK and the arguments length is too long.
This PR resolves it by enabling `--features=compiler_param_file` option for Android NDK, but I'm not sure if all commands in all versions of NDK that should be supported are compatible with this.

Closes #14457.

PiperOrigin-RevId: 448334050