Switch to using the cc_configure in bazel_tools
Previously, with Bzlmod turned on, we would use the cc_configure from rules_cc. That repo is not actively kept in sync with what's actually being used in Bazel right now (which is the version in bazel_tools), so we'll switch to using the version in bazel_tools for now (at least until C++ starlarkification is done).

Related: https://github.com/bazelbuild/bazel-central-registry/pull/90
PiperOrigin-RevId: 450648482