Create mini_tar() for use in our blackbox tests.
This PR uses selected parts of tools/build_defs/pkg.  It is not the way I would have wrote it, but it is based on stuff that has been working for years, so I am disinclined to rewrite it totally. It is solely intended for use bazel's tests, and not distributed as part of the product.

It also contains some vestigial features from the old code that are only half enabled here - like adding symlinks. I want to keep that around in case I need it for the blackbox tests. If not, I can fully delete or enable it after I make that assessment.

Part of #11183

Next PR: Use it in src/test/java/com/google/devtools/build/lib/blackbox/tests/workspace.

Closes #15028.

PiperOrigin-RevId: 450908743
Change-Id: I7c393ab264924ef475e65321bdcd817b02cb0819