Update rules_license to 0.0.3
Updates to the newest release so we can pick up the generic "notice" and "restricted" license types.

This also might fix the bzlmod post-submit. It's hard to tell because my local tests now get past the rules_license problem, but fail on local_config_cc, so there are probably 2 currently broken modules.

Closes #15588.

PiperOrigin-RevId: 452795261
Change-Id: I8338efbb74520bcffbc40fe29a2484409e8ae3ae