Prep build output tests for sh_library as a Starlark rule
When sh_library is implemented as a Starlark rule, the traceback includes the
location of the  sh_library implementation.  Because this would make the test
fragile, this CL adjusts tests in two different ways:

1) checking only the portion of the traceback for the pieces defined in the test.
2) implementing the test using filegroup instead of sh_library.

TESTED:
Ran these tests with and without a patch to exports.bzl that enables
the sh_library as a Starlark rule by default.
PiperOrigin-RevId: 453370399
Change-Id: Ic3ad637e170186abc4c27d2b8682d3dbaa738aff