labels in genquery.scope are no longer configured.
genquery is all about loading phase concepts, so there is no reason why building a `genquery` rule under configuration X should result in the scope being configured.

This is done by introducing a new BuildType.GENQUERY_SCOPE_TYPE_LIST, which represents a dependency, which is followed in the loading phase (so incrementality is not affected) but is NOT followed in the analysis phase (so it does not need to be configured).

RELNOTES:
labels in genquery.scope are no longer configured.
PiperOrigin-RevId: 454035034
Change-Id: I08b2c588565cbed84af9fe33aee121e12892f293