Restructure blob upload code.
The previous blob upload code had a lot of nesting and state shared between Future callbacks that made it hard to follow and modify. There is some fundamental complexity in the problem this code is solving; namely, retrying uploads can resume from the last comitted offset rather than always the beginning of the blob. I hope my rewrite makes the code easier to understand and paves the way for some future behavior modifications I want to make in this area.

Closes #15514.

PiperOrigin-RevId: 454588429
Change-Id: I9bc9c6942534053f19368fe4240dfe074ce8e60d