Workaround for Bazel's Android example app build failure
After bazel commit 0c65082a, the bazel internal Java version was set to 11
universally, which causes issues with bootclasspath for Android builds. In
order to build the example app and other Android integration tests, the java
version must be set to 8 either in the WORKSPACE, build command line, or
javacopts.

RELNOTES: None
PiperOrigin-RevId: 457514737
Change-Id: I802958406b13da97db43223ec25e631c47bc9330