Replace `ProgressLike` marker interface with a boolean method `storeForReplay`.
Makes `Postable` instances not stored by default. In the common case, a post represents some sort of progress event, and storing it in skyframe is a waste of memory.

All existing `Postable`-but-not-`ProgressLike` classes which are posted from a `SkyFunction` (as detected by running presubmit tests) are made to return true from `storeForReplay`. It is possible that some of these were unintentionally stored via the default behavior, but I will audit those in a followup.

PiperOrigin-RevId: 457983032
Change-Id: I6c69b3907c20e3dac362f0c6a9610af17fc4d35d