Track how long it takes to discard the analysis cache by posting `AnalysisCacheClearEvent` to the `EventBus` upon completion.
`BuildInfo.AnalysisCacheClearEvent` is recorded at the beginning of the execution phase. This is different from `BuildInfo.analysis_cache_discarded`, which is recorded at the beginning of the analysis phase due to `--discard_analysis_cache` being specified in the previous build or the build configuration changing in a way that discards it.

PiperOrigin-RevId: 458003892
Change-Id: I8e1b4bb0a68a0b632957c9b9dad1beb529c06663