Deprecate `ALREADY_DECLARED_CHILD_MISSING`.
This should never be tolerated. Modify all places it was passed to `GraphInconsistencyReceiver` to instead crash.

In some places, iterations loop over the requested set of keys instead of the result map so that we can crash if a requested key is not in the map. This necessitates the change in `MemoizingEvaluatorTest`.

As a tangential cleanup, replace `Iterables.concat(entry.getTemporaryDirectDeps())` with `entry.getTemporaryDirectDeps().getAllElementsAsIterable()`. It's strictly better because it doesn't create a new object and it actually returns a `Collection` meaning it has O(1) size.

PiperOrigin-RevId: 458101611
Change-Id: I540479325afe6d4585c1b33624510ccdcc16bd7b