Adds flag `--experimental_worker_strict_flagfiles` that checks if the worker argument list conforms to the spec.
The spec calls for only one flag file, at the end of the argument list. However, the implementation allows for
multiple flag files with startup arguments possible between or after. This has led to subtle bugs.

We could also make the spec conform to the implementation, but there are fewer surprises possible with the spec, and
it makes it much clearer that there is a distinction between startup args and per-request args.

We would like to make it even more explicit, and possibly at the same time allow differing flags for workers and
non-workers (e.g. for JIT/GC tweaking), but that would be a larger change.

PiperOrigin-RevId: 458415814
Change-Id: I7eacfa7e6344348f66b6c83139c3bcd893bdd70b