Replace all usages of Label.parseAbsolute(..., ImmutableMap.of()) in test code with Label.parseCanonical(...)
The former method should eventually be removed and replaced with one of the three new constructors (parseCanonical, parseWithRepoContext, parseWithPackageContext). Calling it with an empty repo mapping is exactly the same as calling parseCanonical, so we replace it with the latter in test code. In production code it's rarely correct to use an empty mapping here, so more audit is required.

PiperOrigin-RevId: 459054214
Change-Id: I42a706519455931271d883640c1d8d2fd4309f6d