Add resource_jars to java_common.compile() for Starlark java_binary
This is needed to pack the timezone data into the class jar. The alternative was to add the artifact to the runtime classpath. This works well with deploy jar, but will fail if one attempts to run a binary class jar without using the stub script. In general, I think it's better to keep 1:1 correspondence with native to begin with.

Keeping this private to builtins for now.

PiperOrigin-RevId: 459279823
Change-Id: Id74cecbd06d8b7b80149fe8c49cdf62daa602744