Add register_{execution_platforms,toolchains} directives to MODULE.bazel files
These are intended to replace the {execution_platforms,toolchains}_to_register attribtues on the module() directive, which are a bit harder to use since they have to be grouped up together near the top.

https://github.com/bazelbuild/bazel/issues/15829

RELNOTES: Added new register_{execution_platforms,toolchains} directives to the MODULE.bazel file, to replace the {execution_platforms,toolchains}_to_register attributes on the module() directive.
PiperOrigin-RevId: 459714400
Change-Id: If8a25ef92ac695a3f012d54d0e189d8ae25d0833