Use '~' instead of '.' as the canonical repo name separator
So far we've been using "module_name.version[.extension_name.internal_repo_name]" as the canonical repo name format. This is problematic since versions can already contain dots, which means that "A.1.0-pre.foo.bar.baz" can be interpreted as "baz generated by extension bar in A@1.0-pre.foo" or "bar.baz generated by extension foo in A@1.0-pre". If we use '~' as the separator, this problem goes away.

PiperOrigin-RevId: 460471660
Change-Id: I4dd4d302c86c0e4dbe0731c75d07cb53bf35deb6