Fix some Python integration tests to pick up rules_cc from distdir_deps.bzl instead of from a very old pinned commit embedded in the code.
The main motivation is that upcoming PRs will need to add rules_license to the workspaces of all integration tests, and using the workspace stanza generation is the right technique for adding repositories.

This also partially addresses https://github.com/bazelbuild/bazel/issues/15878.  It does not bring rules_cc up to date yet, but will make that easier in the future.

RELNOTES: None
PiperOrigin-RevId: 460951078
Change-Id: I70bd1a3cb922f2d796c9924c91666c9b6b168438