Post AnalysisPhaseCompleteEvent for Skymeld.
Many classes, especially the ones collecting metrics around a build, rely on AnalysisPhaseCompleteEvent. For Skymeld, that is problematic since we don't have a clear boundary between analysis/execution.

With this cl, we redefine the meaning of AnalysisPhaseCompleteEvent with --skymeld: it's the moment when all analysis-related work is concluded in the build, and blaze is expected to already be deep in execution work when this moment happens.

This also means that BEP's analysis_phase_time_in_ms [1] keeps its original meaning with Skymeld: it's the wall time duration between the start and the end of analysis. The difference: it's no longer true that total wall time = analysis wall time + execution wall time, as analysis and execution overlap.

To implement this, we introduce AnalysisOperationWatcher, which:
- Has a pre-filled expected set of BuildDriverKeys,
- Listens to TopLevelEntityAnalysisConcludedEvents which include information of the respective BuildDriverKey,
- Removes the BuildDriverKey from the expected set upon receiving the events, and
- Posts an AnalysisPhaseCompleteEvent when the expected set is empty.

The analysis work of a top level target is considered "concluded" when its analysis either fails or succeeds (which also includes work outside of the main Skyframe evaluation, like action conflict checking and compatibility checking).

In case of --nokeep_going and an analysis error occurs, there's no AnalysisPhaseCompleteEvent. This is consistent with the behavior of --noskymeld in the same scenario.

[1] https://github.com/bazelbuild/bazel/blob/1941e2b1b5be8596e984b02fd5dd37e3ed25a81f/src/main/java/com/google/devtools/build/lib/buildeventstream/proto/build_event_stream.proto#L948

PiperOrigin-RevId: 461154740
Change-Id: I731c817280e66a6aad3781cc578c60a708a56ceb