Tighten up ConfiguredTargetKey API by adding fromConfiguredTarget
Constructing a correct ConfiguredTargetKey is perilous as the provided configuration keymust be the post-transition configuration (including self-transitions and trimming) or else issues will arise.

This is part of an effort to audit all creations of a ConfiguredTargetKey to make sure they are compliant with the aforementioned contract.

In this case, a ConfiguredTarget already exists so it is very easy to construct a ConfiguredTargetKey to get it. The new API returns the key already built because that is what all the callsites want (and assures the configuration can't be improperly changed).

PiperOrigin-RevId: 462271525
Change-Id: Id2d3245a618d1f687d909025c897b1ac34369620