Reduce imports processed by python stubs.
There are several import statements in the python wrapper stub used by
`py_binary` and `py_test` rules which are only used when running from a
zip archive.  Processing these imports is a waste of time in the common
case where they aren't going to  be used.  They also might not be
present at all in some stripped-down python environments (e.g. the
Debian `python3-minimal` package), which might be the only python
available on a remote builder when using a hermetic python toolchain
(even when using a hermetic python toolchain, the wrapper is executed
using the system python environment).

Closes https://github.com/bazelbuild/bazel/issues/15571

Closes #15572.

PiperOrigin-RevId: 462362750
Change-Id: I561fde110c8b176e2f616ab9a0469b6c50c6986b