Make blaze ignore --experimental_merged_skyframe_analysis_execution with --nobuild.
When there's no execution phase, running blaze with skymeld doesn't make sense
since there's no overlapping. In that case, we ignore the flag.

Why not throw an error? In the future if someone wants to use aquery/cquery
(which implicitly has --nobuild) while having
--experimental_merged_skyframe_analysis_execution in their bazelrc, it would
add too much friction if an error was thrown.

PiperOrigin-RevId: 462392726
Change-Id: I7c8a379ccb7318cf8c6dee38a67e2a8ee14e9683