BEGIN_PUBLIC Allow StarlarkValue#str to access StarlarkSemantics
This allows us to alter the behavior of `str(some_starlark_value)` depending on flags. See child CL for an example.

This involves adding the StarlarkSemantics argument to a rather long list of call sites and implementors, but it helps in the end to clarify that stringification is really dependent on starlark semantics.
END_PUBLIC

Work towards https://github.com/bazelbuild/bazel/issues/15916

PiperOrigin-RevId: 462564920
Change-Id: I1733424a5657e54cb19f117b8b5dc7af76ca0842