Move link to make document a bit easier to follow.
Move the link to `#accessing-attributes-with-transitions`
from the section about how to define 1:2 transitions into the section
on how to attach an outgoing transition to an attribute, since that's
the section people are most likely to be reading when they need to
see that link.

This got me, leading to me opening https://github.com/bazelbuild/bazel/issues/15892, because I wasn't implementing a 1:2+ transition and hadn't kept reading to the end of the document.

Closes #15895.

PiperOrigin-RevId: 463243990
Change-Id: I77584e2a0e5fd44695d5e7ce5ba23127d6fc278e