Default experimental_inmemory_dotd_files to true
I asked around a bit about this flag and it sounds like there isn't a
downside to flipping this on by default. The upside is that if you
switch between using `--remote_download_minimal` and not, or switch
between a remote execution config that set this flag, but weren't
setting it globally, you would thrash your outputs and rebuild the
world. This was just a subtle gotcha if you were scoping this flag to
one of these specific configurations.

Closes #15203.

PiperOrigin-RevId: 463407383
Change-Id: I4a0a4b564db12094defe3c48c14915682281eef2