Stop Blaze crashes due to transition inputs of values Starlark can't read
Have Starlark.java return a new UnreadableInStarlarkException that can be caught separately as-needed.

Defer construction of the Dict from FunctionTransitionUtil to StarlarkDefinedConfigTransition where the proper mutability is available. Then, can just convert them all and error gracefully if a requested input is not available.

NOTE: There is a slight performance 'adjustment'.
Performance loss: All of the values are converted regardless of if the Starlark implementation actually reads them (versus before where they are converted as they are read). However, the list has already been curated to those that the Starlark impl specifically asked for so this should be fine in practice.
Performance gain: The new implementation avoids converting again if the same entry is asked for twice.
PiperOrigin-RevId: 463515966
Change-Id: Iffeb06ac1f05e4b8c37392d11751d7ef8924521e