Reduce AndroidAssetMerger intermediate outputs
They are not required in the final build.

For our builds:-

This results in 503MB of outputs (assets.zip files) - each of these includes the transitive closure of all assets. To put in perspective we have only 2.5MB of assets.

This is very costly for remote builds with caches especially on slower connections

If resource conflict checking is not enabled (it is not for our build) this will also avoid running AndroidAssetMerger all together saving 896 actions ~3mins total CPU time.

https://github.com/bazelbuild/rules_android/issues/10

Closes #11303.

PiperOrigin-RevId: 463912957
Change-Id: I61b0cdd03ce5bdfd70f17feaf03529a683cddefc