Bootstrapping on Windows requires JDK 11, just like on non-Windows
When I attempted to bootstrap Bazel on Windows by following the
directions, I discovered that JDK 8 (specified in the directions)
did not work, but that JDK 11 (as specified in the directions for
non-Windows system) did work.  So the directions for Windows are
apparently stale.

In order to fix them, we here simply copy the non-Windows JDK version
advice to the Windows section.

Fixes #16004.

Closes #16006.

PiperOrigin-RevId: 464775438
Change-Id: I33de90c9ae2647d3567017033f1e5082ce37cea6