Reduce the scope of the static cache in `StarlarkRuleTransitionProvider`.
The static cache holds on to data that is obsolete after a `clean` command or non-incremental build (i.e. `--notrack_incremental_state`). This isn't a memory leak because of the use of `softValues()`, but it makes memory profiling noisy and also likely causes the garbage collector to work harder than necessary to keep around cache entries that will never be reused.

Instead of the cache being keyed on `(StarlarkDefinedConfigTransition, Rule)`, create one cache per `StarlarkDefinedConfigTransition` instance and make the key just `Rule`. This achieves better scoping and also allows us to switch from `softValues()` to `weakKeys()` with no performance loss - `Rule` already uses identity equality, and this switch allows entries to be collected as soon as the corresponding `Rule` is collected instead of sticking around until there is memory pressure.

PiperOrigin-RevId: 465357707
Change-Id: Iefef87d06d1b879cc1af6f6ac7560804675f5ec6