Buildfix Netty filter on this select default case.
The `filter_netty_dynamic_libs` genrule has no srcs when not Darwin,
Linux, or Windows. In this case, attempting to invoke the genrule breaks
the build because the `cp` command has no input. Instead, only depend on
this genrule output when the platform is Darwin, Linux, or Windows.

There might be a nicer way of doing this, but this works.

Partial commit for third_party/*, see #16065.

Signed-off-by: Sunil Gowroji <sgowroji@google.com>