Prevent the native Apple split transition from setting ios_cpu in splits. Mark the field as deprecated.
Taking this opportunity to tidy up some of the AppleConfiguration code to make it clear that either --ios_multi_cpus or --cpu if it leads with "ios_" is the source of truth for an iOS build.

In marking the --ios_cpu command line option as deprecated, we start alerting users that the option is likely not what they want before outright removing it. This option has been a no-op since 2017 based on past Bazel release notes.

PiperOrigin-RevId: 467228527
Change-Id: If15a1f151424552c5bbd13526f4fdab8e27cb5c2