Split some behavior from `AbstractSkyFunctionEnvironment` to `AbstractSkyFunctionEnvironmentForTesting`.
Frees up the real `SkyFunctionEnvironment` to make further optimizations which diverge from the test versions (will be done in future changes) while still allowing the test versions to work with simple types (e.g. `Map`).

Exception processing logic is moved to a utility method in `SkyFunctionException`.

Also move `SkyFunctionEnvironmentForTesting` to under `src/test/java/`. It was weird that it was referenced from `SkyframeExcecutor` and built into the bazel binary.

PiperOrigin-RevId: 468287053
Change-Id: I315fb96155f59f8b47d0ddbb8c6b083188309572