Introduce graph methods that don't create maps.
Switch to the new methods for callers that don't need map functionality (deduplication or size). No callers of `createIfAbsentBatch` need map functionality, so the map-returning version of that is now gone.

For some callers, I tried to make it obvious that deduplication is unnecessary by changing the declared type of the requested keys to `Set` where it isn't already. Others requested keys come from `GroupedList#getAllElementsAsIterable` representing deps, which although not a `Set`, is guaranteed to have unique elements. If we really wanted to, we could have that return a marker interface like `UniqueIterable`, but I didn't feel strongly enough to do that.

This is expected to reduce garbage and lower peak post-GC heap.

PiperOrigin-RevId: 468606030
Change-Id: I19132e48f69c55114a8bad30f19da0892635f3c3