Control worker lifecycle via `WorkerLifecycleManager`.
We want to control the abusing of resources (e.g. memory) by workers, so we decided to start with total memory limit for all workers. To control this use the option `experimental_total_worker_memoty_limit_mb`.

To do it we use standard `evict()` function of `GenericKeyedObjectPool`. Indeed we use this call twice: first to collect the idle workers, and for the second time to evict the largest idle workers.

PiperOrigin-RevId: 469693732
Change-Id: Ib0e51587f6b1744b128678f6e341964820bb1aec