Update android/incremental_install to python3.
More for #16105

This took way too long to figure out.

I suspect that incremental_install.py is actually sort of brittle,
but it happens to work by luck for most people.
Getting it out of Bazel and in to rules_android would make it eaisier to maintain.
That is beyond the scope of my current effort to reduce @bazel_tools

Closes #16154.

PiperOrigin-RevId: 470974045
Change-Id: I2ac18de62921fecad019acfeb5f017007098c50b