Remove uses of java_toolchain_alias
This seems necessary so that we can start using C++ toolchain resolution.

Without it, rules that are using both C++ and Java toolchain can resolve to a second platform (when there is no crosscompiler available on the first platform) and then JDK binaries for the first platform get executed on the second one.

It's only necessary to use the toolchain directly on rules that also have other toolchains, like java_binary.

However on rules, where only JDK toolchain is needed like java_library, the whole mechanism seems to work because a Java toolchain is always available on the first platform. If that wasn't the case - toolchain resolution in java_library would still select first platform due to the use of alias (and the alias would select second execution platform).

(1 execution platform is matched to each target, based on the toolchains required by the rule)

PiperOrigin-RevId: 472313976
Change-Id: I1905c25e5caa6fb5269f1a9b9ee12bf1b9c58ba9