Preserve a Starlark module's doc string in the module
Required for the new Stardoc rewrite: we plan to retrieve the module from
skyframe at a point when the module's corresponding StarlarkFile and Program no
longer exist, so we can't directly reach into the parsed file's body to look for the
first statement.

As a side effect, preserving the doc string in the module considerably simplifies
the existing Stardoc's renderer logic. (Testing the existing Stardoc renderer
required a small refactor to split out the proto renderer into a separate function.)

PiperOrigin-RevId: 472782957
Change-Id: Ic8214177b48f4adea026babcc9efb16aaabc07a2