Remove files and remap_paths attributes from deprecated pkg_tar.
This version of pkg_tar was deprecated with the introduction of @rules_pkg
and Bazel 5.0.  It was suppossed to be completely removed by 6.0, but
that was delayed because it still seems essential to one integration
test. This CL is one of several that will agressively cut out features
that are not needed for this one test, so that users of Bazel 6.x don't
continue to use this feature.

PiperOrigin-RevId: 473129108
Change-Id: I0f52504fe7826e2dc3545b91dd817e65e5e94d28