Replace objc_archive action with cpp_static_library action
The former is largely redundant.  The only thing objc_archive had was
bespoke -filelist support that generates extra actions that don't seem
to be worth the cost.  Unlike link actions, objc_archive's commandline
does not grow with size of build, so is not likely to exceed OS limit.
There were no such instance internally.  Eventually we plan to use
linker param files instead, which will be a per-target opt in rather
than always on.

PiperOrigin-RevId: 473323168
Change-Id: I94e0bede6071b0e2a7bbeeb92784a718b28bb866