Bazel docs: update links to persistent workers protos
Currently these links are to an older version of the proto files. That older version is missing some of the fields that are documented in this file. Specifically it is missing `cancal`, `verbosity`, and `sandbox_dir`.

The new commit ID in this change is to the most recent change to the proto file on the master branch.

PiperOrigin-RevId: 473841778
Change-Id: I3681b42b38a1492fa1313b58075496948b4f8bff