[repository/downloader] Add support for multiple header values
Headers can generally be specified multiple times, not just once.

This refactoring is prerequisite for adding support for getting credentials from the credential helper.

Note that this does not change the Starlark API or the qualifier for the gRPC remote downloader (both of which need to go through the process for incompatible changes - which should hopefully be pretty straight forward to do as this feature doesn't seem to be used that much AFAICT).

Progress on #15856

Closes #16260.

PiperOrigin-RevId: 474770302
Change-Id: I74620e36481414a41108991bc61321d104a6d39a