Make docgen handle modules with fields that are constructors.
This makes it so a module's Java implementation can specify the specific provider
type a function returns (rather than `ProviderApi`) without crashing the doc
generator _and_ it allows the generated docs to automatically link to the
provider's documentation page.

Previously, the doc generator would crash because it would attempt to
register the same doc twice for the same provider. e.g., `ExecutionInfo` would
be registered when it visited `ExecutionInfoApi`, then again when it
visited `TestingModuleApi.executionInfo`.

PiperOrigin-RevId: 475683849
Change-Id: I54e1da77581a49954584666a5f392d7b8d04b017