[Skymeld] Skip action conflict checking while bubbling error.
Action conflict checking requires iterating the transitive ConfiguredTargets to
build up the action graph. This information will not be available if we're
running with --discard_analysis_cache and the build has progressed past
AnalysisPhaseCompleteEvent. It's also excessive to run conflict checking when
bubbling error up:
1) it doesn't request any SkyKey evaluation, thus doesn't play a role in error
bubbling, and
2) no new action conflict is expected to show up after the initial check in the
original BuildDriverKey evaluation.

PiperOrigin-RevId: 475832913
Change-Id: Ia9d0e67976d1de6e7eca8f292630de7f50a93e99