Fix lld detection on macOS
The MachO version of LLD is being used in production at this point, but
the detection incorrectly detected the linker name on macOS. Since the
detection was using `string.find("lld")` it matched this flag instead
`-debugger-tuning=lldb`. This now verifies that the match is against the
expected format which starts with double quotes.

Fixes https://github.com/bazelbuild/bazel/issues/15126

Closes #15006.

PiperOrigin-RevId: 475890776
Change-Id: I36843ec86f606bb3fb66fea702c901268fb8d750