Clarify wording regarding updating documentation
The policy suggests updating documentation in the same CL as code changes. However, it wasn't clear to me whether this is just generally recommended as a best practice, or if there's some technical reason related to how release notes are generated (e.g., automatically resolving doc links to their state as of certain commits). My guess is that it's the former and I've updated the wording accordingly, but please let me know if I'm wrong.

PiperOrigin-RevId: 476100199
Change-Id: I884427bb5b68c7381e7cf68330382e0984be8469