Spruce up test-only bugreporter crash-handling business
Wrap all unprocessed throwables with an exception instead of just checked ones.
This makes it easier to understand the source of failures due to unchecked
exceptions (in tests), as the maybePropagateUnprocessedThrowableIfInTest call
is typically a bit removed from the code under test (eg in an @After of
BuildIntegrationTestCase).

Alas, this does make it difficult to assert on what unhandled exception was
thrown, so also add a getAndResetLastCrashingThrowableInTest. While we now
have two methods, we also have the ability to assert on checked exceptions, if
we find ourselves so inclined.

Replaced a maybePropagateUnprocessedThrowableIfInTest in
BuildEventServiceIntegrationTest with just throwing the exception. The spirit of
the test is testing the exit code, which the handleCrash accomplishes, after
that point we're mostly in make-believe land since in reality bug-report would
have crashed the thread. Before the maybePropagate would do that by rethrowing
the exception, now just make it explicit, since we're wrapping the exception.

Finally, renamed the methods to be a bit more specific in what they relate to,
I think the previous naming goes back to their introduction 5+ years ago, the
code has moved a lot since then.

PiperOrigin-RevId: 476887897
Change-Id: I70310a08b3f8e0c86ad0e89908a4e50fc7dabfe2