Implement `SkyframeLookupResult` and `SkyframeIterableResult` without temporary data structures.
They now retrieve values by going to the backing maps in `SkyFunctionEnvironment`. This avoids constructing a garbage map/list for every Skyframe request. It also avoids constructing the `ValueOrUntypedException` wrapper.

A behavioral change is that now values requested in a previous request can be retrieved from a later result even if they were not requested. This seems benign as the deps are already properly accounted for, and if it currently happens anywhere, it would have triggered the bug report.

There is really no reason to use `getOrderedValuesAndExceptions` now. It just delegates to `getValuesAndExceptions` and creates an additional `SkyframeIterableResult`. I left a TODO to delete it.

`SimpleSkyframeLookupResult`, `SimpleSkyframeIterableResult`, and `ValueOrUntypedException` are moved to the test source tree.

PiperOrigin-RevId: 477287013
Change-Id: I90262d16a31f056dcdcaecd9ca4cb1c6afe6c5b8