[Skymeld] Make cycle reporting work with Skymeld.
There were a few corners of cycle reporting that didn't quite work with
Skymeld, which showed up in the tests that we ran with the skymeld flag turned
on.

This CL fixes those corners. One difference in behavior between Skymeld and
regular blaze:

In regular blaze:

- We report cycles once per top level SkyKey that's requested via
`SkyframeExecutor`. In a test command e.g. `blaze test //foo`, these SkyKeys
would be `TargetCompletionKey(//foo)`, `TestCompletionKey(//foo)` and other
special artifacts (e.g. `build-info.txt`)
- If there's a cycle while executing the actions `A1` and `A2` of `//foo`, we'd
do cycle reporting twice for the same cycle: once for `TargetCompletionKey`,
once for `TestCompletionKey`.
- Reporting the cycle twice is too spammy, so we'd skip the 2nd reporting and
print "target depends on an already-reported cycle"

In Skymeld: the difference starts from the first layer: we'd only request for
the evaluation of a single `BuildDriverKey(//foo)`. There's no risk of duplicate
reporting in this case.

PiperOrigin-RevId: 477377879
Change-Id: I1e0203c8237d3cccc7d673a1bd089b87733e59e4