Fix up CcLinkingContexts created by objc_library and j2objc_library
Fix them so that they propagate alwayslink and dependent
CcLinkingContexts.  They are not currently used in most cases, so the
bugs may not have surfaced.  Notably, the CcLinkingContext created by
the build actions in objc_library is currently being thrown away, and
a new one rebuilt, which is redundant and inefficient.  We want to fix
the former and switch to that.

PiperOrigin-RevId: 479476261
Change-Id: Id12d40cc73891ea570e8a856a36868973d7ee1a8