Add Objc.static_framework_file to fully link arguments
When you are creating a fully linked static library that directly depends on static frameworks, you ideally want this framework to be linked, just as if it was an imported static library. Previously it would be silently ignored and you would have to distribute the static framework with the fully linked library. If users need the previous behavior they should use `avoid_deps` when creating the fully linked library.

Closes #16366.

PiperOrigin-RevId: 479580008
Change-Id: Ifff0ce7d2a591c4287b40a1afa18b501e269a87b