Ignore params file for cpp compile `action.argv`
In cases where `compiler_param_file` is enabled, the action parameters are stored in a virtual file. When inspected from starlark, e.g. an aspect, this file cannot be read from another action to get its contents unless the action is executed and `materialize_param_files`. This CL returns the complete list of arguments for `action.argv` guarded by a flag.

PiperOrigin-RevId: 480910893
Change-Id: I830d94206ff7612bbb197017429d1cdda1b70a08