Verify that incompatible options are preserved in the exec cfg
Adding incompatible options is error-prone as it is easy to forget to add them to their corresponding `FragmentOption`'s `getHost` method. If this is not done, the flag will have no effect in the exec configuration, which has already caused very surprising, buggy behavior in numerous cases.

This commit adds a test to verify that all non-deprecated incompatible flags:

1. are tagged with the `INCOMPATIBLE_CHANGE` metadata tag;
2. are preserved in the exec configuration.

Fixes #12238
Fixes #16388

Closes #16449.

PiperOrigin-RevId: 481129974
Change-Id: I4d235e131dc4974380a66460b8bea0c3d0113c67