Prefetch inputs before acquiring the resource and sending Executing event.
prefetchInputs could wait for network I/Os to download all the inptus for this local action which might take a long time. It is possiable that local actions acqure resource and then wait for the network I/Os which is a waste.

The prefetch is also moved before sending Executing event so that the UI will display the action is being prepared instead of running. i.e.

```
[Prepa] Linking xxx
```

instead of

```
Linking xxx
```

when prefetching inputs for the action.

PiperOrigin-RevId: 481874944
Change-Id: I5cd9d379d3ddc68ddf444d4c0ebce5f1ae3ef905