Introduce a separate TreeArtifactCompositeFileArtifactValue class.
In https://github.com/bazelbuild/bazel/pull/16283, I will add a new data member to this class, which won't be shared with other uses of FileArtifactValue#createProxy. This is a preliminary no-op CL to make the subsequent change simpler.

As noted in the PR discussion, the getType() method on the proxy object currently returns REGULAR_FILE instead of DIRECTORY. I intend to fix this in a separate CL since it's unclear whether any callers rely on this.

PiperOrigin-RevId: 481913076
Change-Id: I44f721f10230ec782f7a583cdf372972376959c0