Do not require option converters to be public.
Change how we reflectively instantiate the option converters so that Bazel no longer crashes when we try to use a non-public class. Technically, this is overriding the visibility of the class and the constructor, however this is merely a side effect of the fact that annotations do not allow us to use an actual object.

It is not obvious why these classes have to be public and a mistake here results with Bazel failing to start.

PiperOrigin-RevId: 482281892
Change-Id: I064301ce88f83fd55a305385309cfb357e5e054c