Keep the default value of option fields for which parsing failed.
Before, they were set to null and I'm fairly certain it was not intentional: `OptionsParserImpl.getParsedOptions()` only checked whether the option has a value, but not if parsing for it succeeded. Then, if parsing failed, `OptionValueDescription.effectiveValue` remained unset, thus yielding null as the value of the option.

This caused an actual bug with `--remote_download_outputs` . Even more confusing was that due to another, unrelated bug, the stack trace was not printed on the standard error, Bazel simply exited with the "internal error" exit code.

RELNOTES: None.
PiperOrigin-RevId: 482723418
Change-Id: Ia31fb344f48a3fb02aba7bdf9952c61b21672353