Rename "bzl visibility" to "load visibility" in comments and error messages
See discussion in review of unknown commit for rationale. It boils down to the fact that ".bzl visibility" would be the preferred formatting, but this is awkward to write at the beginning of sentences. Both "bzl" and "load" are pieces of terminology the user already knows. The only downside of "load" is that it can be interpreted as a verb rather than as part of a compound noun.

Changes:
- Generally, I substituted mentions of "bzl visibility" and "bzl-visibility" to be "load visibility".
- In some cases I just made it a bare "visibility" where it was obvious, in error messages created by a call to `visibility()`.
- In other cases I specify ".bzl load visibility" for extra clarity where context is weaker.
- One instance of "load-visibility" was de-hyphenated.
- Deleted mention in `visibility()`'s documentation that the feature is experimental.

Instances of "bzlVisibility" occur in the actual java symbols. I didn't change those because it's not user-visible, is extra churn for us, and can be misread more easily ("load" as a verb). I also didn't change the names of user-facing flags because renaming flags is an absolute nightmare and we don't have time before the Bazel cut anyway.

A follow-up will add more user concepts documentation for load visibility.

(This is probably the longest commit message I've written for such a trivial change.)

Work toward #11261.

PiperOrigin-RevId: 482897281
Change-Id: Iabb21e8b00b9cbb5335302e494bc0b0af0fa59af